build: Add rubocop to CI and update ruby version
Adds a rubocop action to our CI pipeline, installs the Airbnb rubocop
gem, and fixes all of the resulting rubocop errors. The new CI action
tests with ruby versions 2.7, 3.0, 3.1, and 3.2.

This commit also changes the .ruby-version to 3.2.1. We intend to
support all of the major ruby versions >= 2.7, but for development
purposes, it is easiest to work with the most modern version because the
developer tooling is more robust and more frequently updated.