build: Remove duplicate CI actions
Previously, we ran CI actions on pushes to any branch and on changes to
a pull request, which resulted in duplicate events being run in pull
requests. This commit changes the GitHub Actions build rules to only run
CI on pushes to the *main* branch and on any changes to a pull request.