readyset: Add `Query` class (#30)
This commit adds a `ReadySet::Query` class that represents a query that
has either been seen by ReadySet, whether it be cached or not. It also
adds methods to this class to serve as an interface to ReadySet's `SHOW
PROXIED QUERIES` and `SHOW CACHES` SQL extensions. Methods to create and
drop caches will be added in a future commit.