docs: Update issue templates (#34)
New feature request and bug report templates for Github issues. 

The feature request was based off the structure of Ethan's Google Doc
describing a problem and a proposed solution.

The abstract from said doc: 
The way we are currently thinking about our Ruby gem makes for a
confusing user experience due to the query status cache and due to the
fact that annotating a method or class doesn’t actually give us a list
of queries to cache at application startup. It may make sense to pursue
an alternative approach where users specify specific queries to cache up
front