Split `Query` into `Query::CachedQuery` and `Query::ProxiedQuery`
The results returned by `SHOW PROXIED QUERIES` and `SHOW CACHES` share
some columns but not others. What's more, some of the columns have
different names across the two commands but contain the same data (e.g.
"proxied query" and "query text" both contain the ReadySet-rewritten
query string).

Previously, we had a single `Query` class that represented both of these
two schemas. It was overloaded, given that it represented rows from
tables with entirely different schemas. This commit splits `Query` into
`Query::CachedQuery` and `Query::ProxiedQuery` to better separate the
concerns. This commit also renames a "seen-but-not-cached" query to be a
"proxied" query, which matches the terminology we use on ReadySet.