Allow users to define ReadySet "query methods" via new `readyset_query` syntax (#48)
This commit adds an extension to `ActiveRecord::Base` that defines a
`readyset_query` class method. This class method provides an easy
interface for users to define named queries within the context of their
models such that the queries will be routed to ReadySet.

NOTE: This PR is stacked on top of #47 and should only be merged after
that PR is merged

Closes #3