readyset: Fix quoting in create/drop cache methods (#66)
This commit fixes the identifier quoting in our Readyset.create_cache!
and Readyset.drop_cache! methods. Specifically, the query name passed to
`DROP CACHE` should always be quoted (since it's an identifier) and if
we are passing an ID to `CREATE CACHE`, we need to quote that as well.
The main idea here is that whenever we pass the a query ID or query name
to `CREATE CACHE` or `DROP CACHE`, it needs to be quoted, and whenever
we pass a query string to `CREATE CACHE` it should not be quoted.

This PR includes a revert of #64 and some additional logic to properly
handle the quoting.

Fixes #63