railtie: Fix connection pools
Previously, the `readyset.connection_pool` railtie was setting up the
connection pool for the gem by calling
`ConnectionHandler.establish_connection`, which was resulting in the
connection pool's database key being set to that of the primary database
instead of "readyset". This commit resolves the issue by relying
directly on `ActiveRecord::Base.establish_connection` to set up the
connection pool. It also adds unit test converage to ensure the
connection pool is set up as expected.