routing: Add automatic failover (#65)
This commit adds an optional automatic failover feature to the gem. If
the number of connection errors to ReadySet exceeds a preconfigured
threshold, ReadySet is considered to be unhealthy, all queries are
routed upstream, and a background task is spawned that periodically
checks ReadySet's health. When ReadySet is determined to be healthy
again, the task is stopped, and queries are routed back to ReadySet.

Closes #45